Measurement

Description

A Haskell module for working with scientific measurements.

Documentation

data Quant a

Data type representing a measurement with a value and unit.

type QuantUnit = [(Integer, Integer)]

Type synonym for [(Integer, Integer)], a list of [(prefix, power)] tuples for the seven fundamental SI units.

coulomb :: QuantUnit

The SI derived unit of electrical charge. 1 C = 1 A s

farad :: QuantUnit

The SI derived unit of electrical capacitance. 1 F = 1 C V^-1

hertz :: QuantUnit

The SI derived unit of frequency. 1 Hz = 1 s^-1

joule :: QuantUnit

The SI derived unit of energy. 1 J = 1 N m

newton :: QuantUnit

The SI derived unit of force. 1 N = 1 kg m s^-2

ohm :: QuantUnit

The SI derived unit of electrical resistance. 1 ohm = 1 V A^-1

pascal :: QuantUnit

The SI derived unit of pressure. 1 Pa = 1 N m^-2

volt :: QuantUnit

The SI derived unit of electrical potential. 1 V = 1 W A^-1

watt :: QuantUnit

The SI derived unit of electrical power. 1 W = 1 J s^-1

plus :: Num a => Quant a -> Quant a -> Quant a

Takes two Quant a's and returns the Quant a corresponding to their sum.

minus :: Num a => Quant a -> Quant a -> Quant a

Takes two Quant a's and returns the Quant a corresponding to their difference.

times :: (Num a, Fractional a) => Quant a -> Quant a -> Quant a

Takes two Quant a's and returns the prefix-free Quant a corresponding to their product.

dividedBy :: (Eq a, Fractional a, Num a) => Quant a -> Quant a -> Quant a

Takes two Quant a's and returns the Quant a corresponding to the first divided by the second.

fromStringToQuant :: (Fractional a, Num a, Read a) => String -> Quant a

Takes a String containing value and unit information and returns a Quant a. Unit prefixes are not supported yet (except for kg).

pureQuant :: Num a => a -> Quant a

Takes a number and returns a minimal Quant with dimensionless units.

shiftPrefix :: (Num a, Fractional a) => Integer -> Quant a -> Quant a

Takes an integer (n) and a Quant a, and returns a Quant a in which the prefix of the unit has been shifted n decimal places. If a shift to an unknown prefix is attempted, Measurement.Nothing is returned. TO DO: deal with shifts of units to powers. Doesn't really work right now.

showFundamental :: (Eq a, Show a) => Quant a -> String

Takes a Quant a and returns a String representation of the measurement with fundamental SI units only.

showPretty :: (Eq a, Show a, Num a, Fractional a) => Quant a -> String

Takes a Quant a and returns a String representation of the measurement, using derived units where applicable.

unitTimes :: QuantUnit -> QuantUnit -> QuantUnit

Takes two QuantUnits and returns the QuantUnit resulting from their multiplication.

unitToThePowerOf :: QuantUnit -> Integer -> QuantUnit

Takes a QuantUnit and an Integer (n) and returns the QuantUnit resulting from the raising of the QuantUnit to the power of n.

unitOver :: QuantUnit -> QuantUnit -> QuantUnit

Takes two QuantUnits and returns the QuantUnit resulting from their division.